refactor(modularization): Phase 5 step 5-S4b — chat_service move (task #26)
Move the largest of the six conversation services (310 LOC of business
logic around chat CRUD + v3 history hydration) to its canonical domain
home, and replace the legacy ``aperag/service/chat_service.py`` with a
thin re-export shim. chat_service is picked before
``chat_collection_service`` / ``bot_service`` because it has zero
cross-service dependencies — it goes straight to ``db_ops`` and the
domain's own Pydantic schemas — so moving it does not force any other
service to move first and does not need a Phase 4 identity rebase.

Rewritten imports (all domain-local, zero G1 legacy aggregate hits):

* ``from aperag.db import models as db_models`` → three explicit
  domain imports:
  - ``Chat`` / ``ChatStatus`` / ``BotType`` from
    ``aperag.domains.conversation.db.models``
  - ``AgentArtifactType`` / ``AgentTurnStatus`` from
    ``aperag.domains.agent_runtime.db.models``
  - The ORM ``Chat`` row is aliased ``ChatRow`` inside the module so
    it does not shadow the Pydantic ``Chat`` response schema.
* ``from aperag.schema import view_models`` → explicit symbol pulls
  from ``aperag.domains.conversation.schemas``
  (``Chat`` / ``ChatDetails`` / ``ChatMessage`` / ``ChatUpdate`` /
  ``Reference``). ``view_models.Reference`` / ``view_models.ChatMessage``
  references in ``_map_reference_item`` / ``_build_v3_chat_history``
  switch accordingly.
* Redis history helper (``RedisChatMessageHistory`` / the
  ``get_async_redis_client`` import) stays in ``aperag.utils.history``
  — not a G1 ban target.

No functional change. The only surface-level adjustment is the
``ChatRow`` alias, which keeps the ``build_chat_response(chat: ChatRow)
-> Chat`` signature readable without two ``Chat`` names colliding in
the same module scope.

Verify:
* ``from aperag.service.chat_service import ChatService,
  chat_service_global`` OK via shim; ``ChatService is
  aperag.domains.conversation.service.chat_service.ChatService``
* ``pytest tests/unit_test/`` 670 passed / 29 skipped / 1 deselected
  (``test_performance_comparison`` is a pre-existing wall-clock flake,
  unrelated to this carve)
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
* G1: ``aperag/domains/conversation/service/chat_service.py`` contains
  zero imports from ``aperag.db.models`` / ``aperag.service.*`` /
  ``aperag.schema.view_models``.
