+++ aperag/service/chat_service.py
@@ -12,299 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import logging
-from typing import Optional
+"""Legacy shim for ``chat_service`` after Phase 5 step 5-S4b.
 
-from sqlalchemy.ext.asyncio import AsyncSession
+The canonical home is
+``aperag.domains.conversation.service.chat_service``; this module
+re-exports ``ChatService`` and the ``chat_service_global`` singleton
+so pre-migration callers (``views/chat.py`` / ``views/bots_v2.py`` /
+``agent_runtime`` lazy imports / ``evaluation_v2.worker``) keep
+resolving the same class objects without a rename sweep. Phase 6
+cleanup removes the shim after every caller has migrated.
+"""
 
-from aperag.db import models as db_models
-from aperag.db.ops import AsyncDatabaseOps, async_db_ops
-from aperag.exceptions import ChatNotFoundException, ResourceNotFoundException, ValidationException
-from aperag.schema import view_models
-from aperag.schema.view_models import Chat, ChatDetails
-from aperag.utils.history import (
-    RedisChatMessageHistory,
-    get_async_redis_client,
+from aperag.domains.conversation.service.chat_service import (  # noqa: F401
+    ChatService,
+    chat_service_global,
 )
 
-logger = logging.getLogger(__name__)
-
-
-def _artifact_type_value(artifact) -> Optional[str]:
-    if not artifact:
-        return None
-    artifact_type = getattr(artifact, "artifact_type", None)
-    return artifact_type.value if hasattr(artifact_type, "value") else artifact_type
-
-
-def _extract_artifact_text(artifact) -> str:
-    if not artifact or not isinstance(getattr(artifact, "payload", None), dict):
-        return ""
-    return artifact.payload.get("text") or artifact.payload.get("content") or artifact.payload.get("message") or ""
-
-
-def _coerce_timestamp(value) -> Optional[float]:
-    return value.timestamp() if value else None
-
-
-def _map_reference_item(item: dict) -> view_models.Reference:
-    return view_models.Reference(
-        score=item.get("score"),
-        text=item.get("snippet") or "",
-        metadata={
-            **(item.get("metadata") or {}),
-            "title": item.get("title"),
-            "source_type": item.get("source_type"),
-            "source_id": item.get("source_id"),
-            "uri": item.get("uri"),
-        },
-    )
-
-
-def _extract_references(artifact) -> list[view_models.Reference]:
-    if not artifact or not isinstance(getattr(artifact, "payload", None), dict):
-        return []
-    items = artifact.payload.get("items")
-    if not isinstance(items, list):
-        return []
-    return [_map_reference_item(item) for item in items if isinstance(item, dict)]
-
-
-class ChatService:
-    """Chat service that handles business logic for chats"""
-
-    def __init__(self, session: AsyncSession = None):
-        # Use global db_ops instance by default, or create custom one with provided session
-        if session is None:
-            self.db_ops = async_db_ops  # Use global instance
-        else:
-            self.db_ops = AsyncDatabaseOps(session)  # Create custom instance for transaction control
-
-    def build_chat_response(self, chat: db_models.Chat) -> view_models.Chat:
-        """Build Chat response object for API return."""
-        return Chat(
-            id=chat.id,
-            title=chat.title,
-            bot_id=chat.bot_id,
-            peer_type=chat.peer_type,
-            peer_id=chat.peer_id,
-            created=chat.gmt_created.isoformat(),
-            updated=chat.gmt_updated.isoformat(),
-        )
-
-    async def _build_v3_chat_history(self, user: str, chat_id: str) -> list[list[view_models.ChatMessage]]:
-        turns = await self.db_ops.query_agent_turns(user, chat_id)
-
-        history: list[list[view_models.ChatMessage]] = []
-        for turn in turns:
-            history.append(
-                [
-                    view_models.ChatMessage(
-                        id=turn.id,
-                        type="message",
-                        role="human",
-                        data=turn.input_text,
-                        timestamp=_coerce_timestamp(turn.gmt_created),
-                    )
-                ]
-            )
-
-            artifacts = await self.db_ops.query_agent_artifacts_by_turn(turn.id)
-            answer_artifact = (
-                next((artifact for artifact in artifacts if artifact.id == turn.answer_artifact_id), None)
-                if turn.answer_artifact_id
-                else None
-            )
-            if not answer_artifact:
-                answer_artifact = next(
-                    (
-                        artifact
-                        for artifact in artifacts
-                        if _artifact_type_value(artifact) == db_models.AgentArtifactType.ANSWER.value
-                    ),
-                    None,
-                )
-
-            reference_artifact = (
-                next((artifact for artifact in artifacts if artifact.id == turn.reference_bundle_artifact_id), None)
-                if turn.reference_bundle_artifact_id
-                else None
-            )
-            if not reference_artifact:
-                reference_artifact = next(
-                    (
-                        artifact
-                        for artifact in artifacts
-                        if _artifact_type_value(artifact) == db_models.AgentArtifactType.REFERENCE_BUNDLE.value
-                    ),
-                    None,
-                )
-
-            answer_text = _extract_artifact_text(answer_artifact)
-            if not answer_text and turn.status in {
-                db_models.AgentTurnStatus.FAILED,
-                db_models.AgentTurnStatus.CANCELLED,
-            }:
-                answer_text = turn.error_message or ""
-
-            ai_parts: list[view_models.ChatMessage] = []
-            if answer_text:
-                ai_parts.append(
-                    view_models.ChatMessage(
-                        id=turn.id,
-                        type="message",
-                        role="ai",
-                        data=answer_text,
-                        timestamp=_coerce_timestamp(turn.gmt_finished) or _coerce_timestamp(turn.gmt_created),
-                    )
-                )
-            else:
-                ai_parts.append(
-                    view_models.ChatMessage(
-                        id=turn.id,
-                        type="start",
-                        role="ai",
-                        data="",
-                        timestamp=_coerce_timestamp(turn.gmt_created),
-                    )
-                )
-
-            references = _extract_references(reference_artifact)
-            if references:
-                ai_parts.append(
-                    view_models.ChatMessage(
-                        id=turn.id,
-                        type="references",
-                        role="ai",
-                        data="",
-                        references=references,
-                        timestamp=_coerce_timestamp(turn.gmt_finished) or _coerce_timestamp(turn.gmt_created),
-                    )
-                )
-
-            history.append(ai_parts)
-
-        return history
-
-    async def create_chat(self, user: str, bot_id: str) -> view_models.Chat:
-        # First check if bot exists
-        bot = await self.db_ops.query_bot(user, bot_id)
-        if bot is None:
-            raise ResourceNotFoundException("Bot", bot_id)
-        if bot.type != db_models.BotType.AGENT:
-            raise ValidationException("Only agent bots are supported")
-
-        # Direct call to repository method, which handles its own transaction
-        chat = await self.db_ops.create_chat(user=user, bot_id=bot_id)
-
-        return self.build_chat_response(chat)
-
-    async def list_chats(
-        self,
-        user: str,
-        bot_id: str,
-        page: int = 1,
-        page_size: int = 50,
-    ):
-        """List chats with pagination, sorting and search capabilities."""
-
-        # Define sort field mapping
-        sort_mapping = {
-            "created": db_models.Chat.gmt_created,
-        }
-
-        # Define search fields mapping
-        search_fields = {"title": db_models.Chat.title}
-
-        async def _execute_paginated_query(session):
-            from sqlalchemy import and_, desc, select
-
-            # Build base query
-            query = select(db_models.Chat).where(
-                and_(
-                    db_models.Chat.user == user,
-                    db_models.Chat.bot_id == bot_id,
-                    db_models.Chat.status != db_models.ChatStatus.DELETED,
-                )
-            )
-
-            # Build query parameters
-            from aperag.utils.pagination import ListParams, PaginationHelper, PaginationParams, SortParams
-
-            params = ListParams(
-                pagination=PaginationParams(page=page, page_size=page_size),
-                sort=SortParams(sort_by="created", sort_order="desc"),
-            )
-
-            # Use pagination helper
-            items, total = await PaginationHelper.paginate_query(
-                query=query,
-                session=session,
-                params=params,
-                sort_mapping=sort_mapping,
-                search_fields=search_fields,
-                default_sort=desc(db_models.Chat.gmt_created),
-            )
-
-            # Build chat responses
-            chat_responses = []
-            for chat in items:
-                chat_responses.append(self.build_chat_response(chat))
-
-            return PaginationHelper.build_response(items=chat_responses, total=total, page=page, page_size=page_size)
-
-        return await self.db_ops._execute_query(_execute_paginated_query)
-
-    async def get_chat(self, user: str, bot_id: str, chat_id: str) -> view_models.ChatDetails:
-        chat = await self.db_ops.query_chat(user, bot_id, chat_id)
-        if chat is None:
-            raise ChatNotFoundException(chat_id)
-
-        messages = await self._build_v3_chat_history(user, chat_id)
-
-        # Build response object
-        chat_obj = self.build_chat_response(chat)
-        return ChatDetails(**chat_obj.model_dump(), history=messages)
-
-    async def update_chat(
-        self, user: str, bot_id: str, chat_id: str, chat_in: view_models.ChatUpdate
-    ) -> view_models.Chat:
-        # First check if chat exists
-        chat = await self.db_ops.query_chat(user, bot_id, chat_id)
-        if chat is None:
-            raise ChatNotFoundException(chat_id)
-
-        # Direct call to repository method, which handles its own transaction
-        updated_chat = await self.db_ops.update_chat_by_id(user, bot_id, chat_id, chat_in.title)
-
-        if not updated_chat:
-            raise ChatNotFoundException(chat_id)
-
-        return self.build_chat_response(updated_chat)
-
-    async def delete_chat(self, user: str, bot_id: str, chat_id: str) -> Optional[view_models.Chat]:
-        """Delete chat by ID (idempotent operation)
-
-        Returns the deleted chat or None if already deleted/not found
-        """
-        # Check if chat exists - if not, silently succeed (idempotent)
-        chat = await self.db_ops.query_chat(user, bot_id, chat_id)
-        if chat is None:
-            return None
-
-        # Direct call to repository method, which handles its own transaction
-        deleted_chat = await self.db_ops.delete_chat_by_id(user, bot_id, chat_id)
-
-        if deleted_chat:
-            # Clear chat history from Redis
-            history = RedisChatMessageHistory(chat_id, redis_client=get_async_redis_client())
-            await history.clear()
-
-            return self.build_chat_response(deleted_chat)
-
-        return None
-
-
-# Create a global service instance for easy access
-# This uses the global db_ops instance and doesn't require session management in views
-chat_service_global = ChatService()
+__all__ = [
+    "ChatService",
+    "chat_service_global",
+]
